+++ wait-for-the-element.d.ts
@@ -26,28 +26,25 @@ export interface ElementSearchOptions
  * Example usage:
  *
  * ``` js
- * let element;
- *
- * try
+ * const element = await waitForTheElement('.target', {
  * {
- *     element = await waitForTheElement('.target', {
- *         timeout : 5000
- *     });
- * }
- * catch (error)
+ *     timeout : 5000
+ * });
+ *
+ * if (element === null)
  * {
  *     // After 5 seconds, a matching element still doesn't exist.
  * }
  * ```
  *
  * @param selector The selector that an element must match. This can be any selector that `document.querySelector()` supports.
- * @param options  The options that define how elements should be searched for.
+ * @param options The options that define how elements should be searched for.
  *
  * @typeParam E The type of the element you expect to be found.
  *
- * @returns A promise that will resolve with the first element matching `selector`, or reject with an error if a matching element is not found.
+ * @returns A promise that will resolve with the first element matching the `selector`, or resolve with `null` if a matching element is not found.
  */
-export function waitForTheElement<E extends Element> (selector : string, options? : ElementSearchOptions) : Promise<E>;
+export function waitForTheElement<E extends Element> (selector : string, options? : ElementSearchOptions) : Promise<E | null>;
 
 /**
  * Waits for all elements to stop matching a provided selector.
@@ -55,21 +52,19 @@ export function waitForTheElement<E extends Element> (selector : string, options
  * Example usage:
  *
  * ``` js
- * try
- * {
- *     await waitForTheElementToDisappear('.target', {
- *         timeout : 5000
- *     });
- * }
- * catch (error)
+ * const hidden = await waitForTheElementToDisappear('.target', {
+ *     timeout : 5000
+ * });
+ *
+ * if (!hidden)
  * {
  *     // After 5 seconds, a matching element still exists.
  * }
  * ```
  *
  * @param selector The selector that elements must not match. This can be any selector that `document.querySelector()` supports.
- * @param options  Options that define how elements should be searched for.
+ * @param options Options that define how elements should be searched for.
  *
- * @returns A promise that will resolve, or reject with an error if elements are still found to be matching.
+ * @returns A promise that will resolve with `false` if elements are still found to be matching the `selector`, otherwise it will resolve with `true`.
  */
-export function waitForTheElementToDisappear(selector : string, options? : ElementSearchOptions) : Promise<void>;
+export function waitForTheElementToDisappear(selector : string, options? : ElementSearchOptions) : Promise<boolean>;